+++ src/typed_settings/click_utils.py
@@ -27,6 +27,7 @@ from typing import (
 import attrs
 import click
 from attr._make import _Nothing as NothingType
+from click.core import ParameterSource
 
 from . import _core, dict_utils
 from .attrs import CLICK_KEY, METADATA_KEY, _SecretRepr
@@ -244,14 +245,17 @@ def _get_wrapper(
                 ctx.obj = {}
             meta = LoaderMeta("Command line args")
             cli_options = ctx.obj.get(CTX_KEY, {})
+            cli_settings: MergedSettings = {}
             for option in state.options:
                 path = option.path
                 if path in cli_options:  # pragma: no cover
                     # "path" *should* always be in "cli_options", b/c we *currently*
                     # generate CLI options for all options.  But let's stay safe here
                     # in case the behavior changes in the future.
-                    merged_settings[path] = LoadedValue(cli_options[path], meta)
-            settings = _core.convert(merged_settings, state)
+                    cli_settings[path] = LoadedValue(cli_options[path], meta)
+                else:
+                    cli_settings[path] = merged_settings[path]
+            settings = _core.convert(cli_settings, state)
             if argname:
                 ctx_key = argname
                 kwargs = {argname: settings, **kwargs}  # type: ignore
@@ -677,10 +681,22 @@ def _make_callback(
         if user_callback is not None:
             value = user_callback(ctx, param, value)
 
-        if ctx.obj is None:
-            ctx.obj = {}
-        settings = ctx.obj.setdefault(CTX_KEY, {})
-        settings[path] = value
+        param_source = ctx.get_parameter_source(param.name or "")
+        param_source_default = param_source in {
+            ParameterSource.DEFAULT,
+            ParameterSource.DEFAULT_MAP,
+        }
+        param_source_user = param_source in {
+            ParameterSource.COMMANDLINE,
+            ParameterSource.PROMPT,
+        }
+        if param_source_user or (param_source_default and value in (None, (), {})):
+            # Don't add the value to the settings if it is a default.
+            # If we did, it would override the original LoaderMeta.
+            if ctx.obj is None:
+                ctx.obj = {}
+            settings = ctx.obj.setdefault(CTX_KEY, {})
+            settings[path] = value
         return value
 
     return cb